Repair broken test submodule on Windows
Hi Alessandro (@aless80),

The "git clone" command is broken on Windows due to an invalid filename in the `test` submodule. The command mentioned in the [main README](https://github.com/doconce/doconce#comprehensive-installation) does not work for me:

`git clone --recurse-submodules git@github.com:doconce/doconce.git`

A modified version (`git clone --recurse-submodules https://github.com/doconce/doconce.git`) starts the clone process, but then fails on a filename that is invalid on Windows:

![invalid-path-on-windows](https://user-images.githubusercontent.com/11787866/198839143-695fa0d5-cb82-47ad-a705-046d05d62657.png)

The `doc` submodule completes correctly, but the `test` submodule fails on the [invalid_papers-20210503-15:05:32.pub](https://github.com/doconce/doconce_tests/blob/main/invalid_papers-20210503-15:05:32.pub) filename since the colon is not valid in a Windows filename.

I [checked](https://github.com/search?q=org%3Adoconce+invalid_papers&type=code) to see where this filename was being used in the **DocOnce** repos and the main use seems to be in the _doconce/publish_ repo:

![invalid-papers-in-doconce-publish-repo](https://user-images.githubusercontent.com/11787866/198849478-7f410f5f-c506-467f-84bc-2a7680a178ce.png)

The _[tests.py (Line 25 to 27)](https://github.com/doconce/publish/blob/master/tests/tests.py#L25-L27)_ test seems to be only usage of the _invalid_papers_ in these repos. Since the date / time is being stripped from the filename as part of the test, then changing this portion of the filename should not affect the test.

If, in fact, the _invalid_papers_ are used in other ways in this repo, and that replacing the colons with dashes in the filename causes other part of the repo to fail, then please ignore this PR.

Kind Regards,
Liam